+++ src/app/features/baropot/components/main/BaropotBanner.tsx
@@ -1,0 +1,17 @@
+export default function BaropotBanner() {
+  return (
+    <div className="mx-4 mt-4 rounded-xl bg-gradient-to-r from-[#1C4E80] to-[#2563eb] p-6 text-white">
+      <div className="flex items-center justify-between">
+        <div>
+          <h2 className="mb-2 text-xl font-bold">⚡ 바로팟이란?</h2>
+          <p className="text-sm opacity-90">
+            혼자 가기 부담스러운 맛집,
+            <br />
+            지금 바로 함께할 사람을 찾아보세요!
+          </p>
+        </div>
+        <div className="text-4xl">🍽️</div>
+      </div>
+    </div>
+  );
+}
